polimorfismo e regras de negócio

+++ SistemaBancario/src/sistemabancario/model/ContaI.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package sistemabancario.model;
 
 public interface ContaI {
@@ -10,7 +5,7 @@ public interface ContaI {
     public boolean deposita(double valor);
     public boolean saca(double valor);
     public Cliente getDono();
-    public int getNumero();
+    public long getNumero();
     public double getSaldo();
     public void remunera();
